Move MapUpdateTask to its own file
