+++ mapcache/src/main/java/mil/nga/mapcache/GeoPackageMapFragment.java
@@ -8,8 +8,6 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
 import android.content.pm.PackageManager;
-import android.graphics.Color;
-import android.graphics.Paint;
 import android.graphics.Point;
 import android.hardware.SensorEvent;
 import android.location.Location;
@@ -87,8 +85,6 @@ import com.google.android.gms.maps.model.Polygon;
 import com.google.android.gms.maps.model.PolygonOptions;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
-import com.google.android.gms.maps.model.TileOverlayOptions;
-import com.google.android.gms.maps.model.TileProvider;
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.textfield.TextInputEditText;
@@ -101,14 +97,10 @@ import java.lang.reflect.Method;
 import java.sql.SQLException;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -118,13 +110,9 @@ import mil.nga.geopackage.GeoPackageException;
 import mil.nga.geopackage.contents.Contents;
 import mil.nga.geopackage.contents.ContentsDao;
 import mil.nga.geopackage.db.GeoPackageDataType;
-import mil.nga.geopackage.extension.nga.link.FeatureTileTableLinker;
-import mil.nga.geopackage.extension.nga.scale.TileScaling;
-import mil.nga.geopackage.extension.nga.scale.TileTableScaling;
 import mil.nga.geopackage.extension.schema.SchemaExtension;
 import mil.nga.geopackage.extension.schema.columns.DataColumns;
 import mil.nga.geopackage.extension.schema.columns.DataColumnsDao;
-import mil.nga.geopackage.features.columns.GeometryColumns;
 import mil.nga.geopackage.features.index.FeatureIndexListResults;
 import mil.nga.geopackage.features.index.FeatureIndexManager;
 import mil.nga.geopackage.features.index.FeatureIndexResults;
@@ -145,24 +133,12 @@ import mil.nga.geopackage.map.geom.GoogleMapShapeType;
 import mil.nga.geopackage.map.geom.PolygonHoleMarkers;
 import mil.nga.geopackage.map.geom.ShapeMarkers;
 import mil.nga.geopackage.map.geom.ShapeWithChildrenMarkers;
-import mil.nga.geopackage.map.tiles.overlay.BoundedOverlay;
-import mil.nga.geopackage.map.tiles.overlay.FeatureOverlay;
 import mil.nga.geopackage.map.tiles.overlay.FeatureOverlayQuery;
-import mil.nga.geopackage.map.tiles.overlay.GeoPackageOverlayFactory;
-import mil.nga.geopackage.srs.SpatialReferenceSystem;
 import mil.nga.geopackage.tiles.TileBoundingBoxUtils;
-import mil.nga.geopackage.tiles.features.DefaultFeatureTiles;
-import mil.nga.geopackage.tiles.features.FeatureTiles;
-import mil.nga.geopackage.tiles.features.custom.NumberFeaturesTile;
-import mil.nga.geopackage.tiles.matrixset.TileMatrixSet;
-import mil.nga.geopackage.tiles.user.TileDao;
 import mil.nga.mapcache.data.GeoPackageDatabase;
 import mil.nga.mapcache.data.GeoPackageDatabases;
-import mil.nga.mapcache.data.GeoPackageFeatureOverlayTable;
-import mil.nga.mapcache.data.GeoPackageFeatureTable;
 import mil.nga.mapcache.data.GeoPackageTable;
 import mil.nga.mapcache.data.GeoPackageTableType;
-import mil.nga.mapcache.data.GeoPackageTileTable;
 import mil.nga.mapcache.data.MarkerFeature;
 import mil.nga.mapcache.indexer.IIndexerTask;
 import mil.nga.mapcache.listeners.DetailActionListener;
@@ -181,7 +157,6 @@ import mil.nga.mapcache.preferences.GridType;
 import mil.nga.mapcache.preferences.PreferencesActivity;
 import mil.nga.mapcache.repository.GeoPackageModifier;
 import mil.nga.mapcache.sensors.SensorHandler;
-import mil.nga.mapcache.utils.ProjUtils;
 import mil.nga.mapcache.utils.SwipeController;
 import mil.nga.mapcache.utils.ViewAnimation;
 import mil.nga.mapcache.view.GeoPackageAdapter;
@@ -472,11 +447,6 @@ public class GeoPackageMapFragment extends Fragment implements
      * Edit clear button
      */
     private ImageButton editClearPolygonHolesButton;
-
-    /**
-     * List of Feature Overlay Queries for querying tile overlay clicks
-     */
-    private final List<FeatureOverlayQuery> featureOverlayQueries = new ArrayList<>();
 
     /**
      * Intent activity request code when choosing a file
@@ -3304,7 +3274,7 @@ public class GeoPackageMapFragment extends Fragment implements
             model.setFeaturesBoundingBox(null);
             model.setTilesBoundingBox(null);
             model.setFeatureOverlayTiles(false);
-            featureOverlayQueries.clear();
+            model.getFeatureOverlayQueries().clear();
             model.getFeatureShapes().clear();
             model.getMarkerIds().clear();
             int maxFeatures = getMaxFeatures();
@@ -3322,7 +3292,7 @@ public class GeoPackageMapFragment extends Fragment implements
                     if (updateFeaturesTask != null) {
                         updateFeaturesTask.cancel();
                     }
-                    updateTask = new MapUpdateTask();
+                    updateTask = new MapUpdateTask(getActivity(), map, basemapApplier, model, geoPackageViewModel);
                     localUpdateTask = updateTask;
                 } finally {
                     updateLock.unlock();
@@ -3330,323 +3300,6 @@ public class GeoPackageMapFragment extends Fragment implements
 
                 localUpdateTask.execute(false, maxFeatures, mapViewBoundingBox, toleranceDistance, filter);
             });
-        }
-    }
-
-    /**
-     * Update the map in the background
-     */
-    private class MapUpdateTask extends AsyncTask<Object, Void, Void> {
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected Void doInBackground(Object... params) {
-            boolean zoom = (Boolean) params[0];
-            int maxFeatures = (Integer) params[1];
-            BoundingBox mapViewBoundingBox = (BoundingBox) params[2];
-            double toleranceDistance = (Double) params[3];
-            boolean filter = (Boolean) params[4];
-            update(this, zoom, maxFeatures, mapViewBoundingBox, toleranceDistance, filter);
-            if (getActivity() != null) {
-                getActivity().runOnUiThread(() -> basemapApplier.applyBasemaps(map));
-            }
-            return null;
-        }
-
-    }
-
-    /**
-     * Update the map
-     *
-     * @param zoom               The current zoom level.
-     * @param task               The update task.
-     * @param maxFeatures        The total number of max features allowed on map.
-     * @param mapViewBoundingBox The bounding box of the current view of the map.
-     * @param toleranceDistance  Used to simplify any geometries being drawn on map.
-     * @param filter             The filter if any.
-     */
-    private void update(MapUpdateTask task, boolean zoom, final int maxFeatures, BoundingBox mapViewBoundingBox, double toleranceDistance, boolean filter) {
-
-        if (model.getActive() != null) {
-
-            // Open active GeoPackages and create feature DAOS, display tiles and feature tiles
-            List<GeoPackageDatabase> activeDatabases = new ArrayList<>(model.getActive().getDatabases());
-            for (GeoPackageDatabase database : activeDatabases) {
-
-                if (task.isCancelled()) {
-                    break;
-                }
-
-                try {
-                    GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(database.getDatabase());
-
-                    if (geoPackage != null) {
-
-                        Set<String> featureTableDaos = new HashSet<>();
-                        Collection<GeoPackageFeatureTable> features = database.getFeatures();
-                        if (!features.isEmpty()) {
-                            for (GeoPackageFeatureTable featureTable : features) {
-                                featureTableDaos.add(featureTable.getName());
-                            }
-                        }
-
-                        for (GeoPackageFeatureOverlayTable featureOverlay : database.getFeatureOverlays()) {
-                            if (featureOverlay.isActive()) {
-                                featureTableDaos.add(featureOverlay.getFeatureTable());
-                            }
-                        }
-
-                        if (!featureTableDaos.isEmpty()) {
-                            Map<String, FeatureDao> databaseFeatureDaos = new HashMap<>();
-                            model.getFeatureDaos().put(database.getDatabase(), databaseFeatureDaos);
-                            for (String featureTable : featureTableDaos) {
-
-                                if (task.isCancelled()) {
-                                    break;
-                                }
-
-                                FeatureDao featureDao = geoPackage.getFeatureDao(featureTable);
-                                databaseFeatureDaos.put(featureTable, featureDao);
-                            }
-                        }
-
-                        // Display the tiles
-                        for (GeoPackageTileTable tiles : database.getTiles()) {
-                            if (task.isCancelled()) {
-                                break;
-                            }
-                            try {
-                                displayTiles(tiles);
-                            } catch (Exception e) {
-                                Log.e(GeoPackageMapFragment.class.getSimpleName(),
-                                        e.getMessage());
-                            }
-                        }
-
-                        // Display the feature tiles
-                        for (GeoPackageFeatureOverlayTable featureOverlay : database.getFeatureOverlays()) {
-                            if (task.isCancelled()) {
-                                break;
-                            }
-                            if (featureOverlay.isActive()) {
-                                try {
-                                    displayFeatureTiles(featureOverlay);
-                                } catch (Exception e) {
-                                    Log.e(GeoPackageMapFragment.class.getSimpleName(),
-                                            e.getMessage());
-                                }
-                            }
-                        }
-
-                    } else {
-                        model.getActive().removeDatabase(database.getDatabase(), false);
-                    }
-                } catch (Exception e) {
-                    Log.e(GeoPackageMapFragment.class.getSimpleName(), "Error opening geopackage: " + database.getDatabase(), e);
-                }
-            }
-
-            // Add features
-            if (!task.isCancelled()) {
-                updateLock.lock();
-                try {
-                    if (updateFeaturesTask != null) {
-                        updateFeaturesTask.cancel();
-                    }
-                    updateFeaturesTask = new MapFeaturesUpdateTask(getActivity(), map, model, geoPackageViewModel);
-                    updateFeaturesTask.execute(maxFeatures, mapViewBoundingBox, toleranceDistance, filter);
-                } finally {
-                    updateLock.unlock();
-                }
-            }
-
-        }
-
-    }
-
-    /**
-     * Display tiles
-     *
-     * @param tiles The tiles to display.
-     */
-    private void displayTiles(GeoPackageTileTable tiles) {
-
-        GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(tiles.getDatabase());
-
-        TileDao tileDao = geoPackage.getTileDao(tiles.getName());
-
-        TileTableScaling tileTableScaling = new TileTableScaling(geoPackage, tileDao);
-        TileScaling tileScaling = tileTableScaling.get();
-
-        BoundedOverlay overlay = GeoPackageOverlayFactory
-                .getBoundedOverlay(tileDao, getResources().getDisplayMetrics().density, tileScaling);
-
-        TileMatrixSet tileMatrixSet = tileDao.getTileMatrixSet();
-
-        FeatureTileTableLinker linker = new FeatureTileTableLinker(geoPackage);
-        List<FeatureDao> featureDaos = linker.getFeatureDaosForTileTable(tileDao.getTableName());
-
-        if (getActivity() != null) {
-            for (FeatureDao featureDao : featureDaos) {
-
-                // Create the feature tiles
-                FeatureTiles featureTiles = new DefaultFeatureTiles(getActivity(), geoPackage, featureDao,
-                        getResources().getDisplayMetrics().density);
-
-                model.setFeatureOverlayTiles(true);
-
-                // Add the feature overlay query
-                FeatureOverlayQuery featureOverlayQuery = new FeatureOverlayQuery(getActivity(), overlay, featureTiles);
-                featureOverlayQuery.calculateStylePixelBounds();
-                featureOverlayQueries.add(featureOverlayQuery);
-            }
-        }
-
-        // Set the tiles index to be -2 of it is behind features and tiles drawn from features
-        int zIndex = -2;
-
-        // If these tiles are linked to features, set the zIndex to -1 so they are placed before imagery tiles
-        if (!featureDaos.isEmpty()) {
-            zIndex = -1;
-        }
-
-        BoundingBox displayBoundingBox = tileMatrixSet.getBoundingBox();
-        Contents contents = tileMatrixSet.getContents();
-        BoundingBox contentsBoundingBox = contents.getBoundingBox();
-        if (contentsBoundingBox != null) {
-            ProjectionTransform transform = contents.getSrs().getProjection().getTransformation(tileMatrixSet.getSrs().getProjection());
-            BoundingBox transformedContentsBoundingBox = contentsBoundingBox;
-            if (!transform.isSameProjection()) {
-                transformedContentsBoundingBox = transformedContentsBoundingBox.transform(transform);
-            }
-            displayBoundingBox = displayBoundingBox.overlap(transformedContentsBoundingBox);
-        }
-
-        displayTiles(overlay, displayBoundingBox, tileMatrixSet.getSrs(), zIndex, null);
-    }
-
-    /**
-     * Display feature tiles
-     *
-     * @param featureOverlayTable The overlay table to display.
-     */
-    private void displayFeatureTiles(GeoPackageFeatureOverlayTable featureOverlayTable) {
-
-        GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(featureOverlayTable.getDatabase());
-        Map<String, FeatureDao> daos = model.getFeatureDaos().get(featureOverlayTable.getDatabase());
-        if (daos != null && getActivity() != null) {
-            FeatureDao featureDao = daos.get(featureOverlayTable.getFeatureTable());
-
-            BoundingBox boundingBox = new BoundingBox(featureOverlayTable.getMinLon(),
-                    featureOverlayTable.getMinLat(), featureOverlayTable.getMaxLon(), featureOverlayTable.getMaxLat());
-
-            // Load tiles
-            FeatureTiles featureTiles = new DefaultFeatureTiles(getActivity(), geoPackage, featureDao,
-                    getResources().getDisplayMetrics().density);
-            if (featureOverlayTable.isIgnoreGeoPackageStyles()) {
-                featureTiles.ignoreFeatureTableStyles();
-            }
-
-            featureTiles.setMaxFeaturesPerTile(featureOverlayTable.getMaxFeaturesPerTile());
-            if (featureOverlayTable.getMaxFeaturesPerTile() != null) {
-                featureTiles.setMaxFeaturesTileDraw(new NumberFeaturesTile(getActivity()));
-            }
-
-            Paint pointPaint = featureTiles.getPointPaint();
-            pointPaint.setColor(Color.parseColor(featureOverlayTable.getPointColor()));
-
-            pointPaint.setAlpha(featureOverlayTable.getPointAlpha());
-            featureTiles.setPointRadius(featureOverlayTable.getPointRadius());
-
-            Paint linePaint = featureTiles.getLinePaintCopy();
-            linePaint.setColor(Color.parseColor(featureOverlayTable.getLineColor()));
-
-            linePaint.setAlpha(featureOverlayTable.getLineAlpha());
-            linePaint.setStrokeWidth(featureOverlayTable.getLineStrokeWidth());
-            featureTiles.setLinePaint(linePaint);
-
-            Paint polygonPaint = featureTiles.getPolygonPaintCopy();
-            polygonPaint.setColor(Color.parseColor(featureOverlayTable.getPolygonColor()));
-
-            polygonPaint.setAlpha(featureOverlayTable.getPolygonAlpha());
-            polygonPaint.setStrokeWidth(featureOverlayTable.getPolygonStrokeWidth());
-            featureTiles.setPolygonPaint(polygonPaint);
-
-            featureTiles.setFillPolygon(featureOverlayTable.isPolygonFill());
-            if (featureTiles.isFillPolygon()) {
-                Paint polygonFillPaint = featureTiles.getPolygonFillPaintCopy();
-                polygonFillPaint.setColor(Color.parseColor(featureOverlayTable.getPolygonFillColor()));
-
-                polygonFillPaint.setAlpha(featureOverlayTable.getPolygonFillAlpha());
-                featureTiles.setPolygonFillPaint(polygonFillPaint);
-            }
-
-            featureTiles.calculateDrawOverlap();
-
-            FeatureOverlay featureOverlay = new FeatureOverlay(featureTiles);
-            featureOverlay.setBoundingBox(boundingBox, ProjectionFactory.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM));
-            featureOverlay.setMinZoom(featureOverlayTable.getMinZoom());
-            featureOverlay.setMaxZoom(featureOverlayTable.getMaxZoom());
-
-            // Get the tile linked overlay
-            BoundedOverlay overlay = GeoPackageOverlayFactory.getLinkedFeatureOverlay(featureOverlay, geoPackage);
-
-            if (featureDao != null) {
-                GeometryColumns geometryColumns = featureDao.getGeometryColumns();
-                Contents contents = geometryColumns.getContents();
-
-                GeoPackageUtils.prepareFeatureTiles(featureTiles);
-
-                model.setFeatureOverlayTiles(true);
-
-                FeatureOverlayQuery featureOverlayQuery = new FeatureOverlayQuery(getActivity(), overlay, featureTiles);
-                featureOverlayQuery.calculateStylePixelBounds();
-                featureOverlayQueries.add(featureOverlayQuery);
-
-                displayTiles(overlay, contents.getBoundingBox(), contents.getSrs(), -1, boundingBox);
-            }
-        }
-    }
-
-    /**
-     * Display tiles
-     *
-     * @param overlay              The tile overlay.
-     * @param dataBoundingBox      The bounding box of the data.
-     * @param srs                  The spatial reference system of the tiles.
-     * @param zIndex               The zoom level.
-     * @param specifiedBoundingBox The specified bounding box.
-     */
-    private void displayTiles(TileProvider overlay, BoundingBox dataBoundingBox, SpatialReferenceSystem srs, int zIndex, BoundingBox specifiedBoundingBox) {
-
-        final TileOverlayOptions overlayOptions = new TileOverlayOptions();
-        overlayOptions.tileProvider(overlay);
-        overlayOptions.zIndex(zIndex);
-
-        BoundingBox boundingBox = dataBoundingBox;
-        if (boundingBox != null) {
-            boundingBox = ProjUtils.getInstance().transformBoundingBoxToWgs84(boundingBox, srs);
-        } else {
-            boundingBox = new BoundingBox(-ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH,
-                    ProjectionConstants.WEB_MERCATOR_MIN_LAT_RANGE,
-                    ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH,
-                    ProjectionConstants.WEB_MERCATOR_MAX_LAT_RANGE);
-        }
-
-        if (specifiedBoundingBox != null) {
-            boundingBox = boundingBox.overlap(specifiedBoundingBox);
-        }
-
-        if (model.getTilesBoundingBox() == null) {
-            model.setTilesBoundingBox(boundingBox);
-        } else {
-            model.setTilesBoundingBox(model.getTilesBoundingBox().union(boundingBox));
-        }
-
-        if (getActivity() != null) {
-            getActivity().runOnUiThread(() -> map.addTileOverlay(overlayOptions));
         }
     }
 
@@ -4054,8 +3707,8 @@ public class GeoPackageMapFragment extends Fragment implements
 
             StringBuilder clickMessage = new StringBuilder();
 
-            if (!featureOverlayQueries.isEmpty()) {
-                for (FeatureOverlayQuery query : featureOverlayQueries) {
+            if (!model.getFeatureOverlayQueries().isEmpty()) {
+                for (FeatureOverlayQuery query : model.getFeatureOverlayQueries()) {
                     String message = query.buildMapClickMessage(point, view, map);
                     if (message != null) {
                         if (clickMessage.length() > 0) {
